# optimize `in(::Char, ::String)`