This is approximately 2x faster for finding ascii characters in strings than before.

```
s = "chalet"
#old
julia> @btime in('\0', $(Ref(s))[])
  10.470 ns (0 allocations: 0 bytes)
false
#new
julia> @btime in('t', $(Ref(s))[])
  4.470 ns (0 allocations: 0 bytes)
true
julia> s = "h"^1000*'t'
#old
julia> @btime in('\0', $(Ref(s))[])
  17.184 ns (0 allocations: 0 bytes)
false
#new
julia> @btime _in('t', $(Ref(s))[])
  9.469 ns (0 allocations: 0 bytes)
true
```
closes https://github.com/JuliaLang/julia/issues/43980