@stevengj 2022-01-30
Might be better to search (with `memchr`) for the first byte of the UTF8 encoding:
```jl
firstbyte = ((bitcast(UInt32, 'x') & 0xff000000) >> 24) % UInt8
```
If it isn't present, return `false`.  If it *is* present and `firstbyte ≤ 0x7f`, return `true`.  Otherwise, call `findnext` starting at the index where `firstbyte` was found.

This way, even the non-ASCII case is accelerated, and even the ASCII case it might be faster because you only do the `isascii` check if it is found.