@jakobnissen 2022-01-31
Perhaps put the memchr in a while loop so it's fast even if the first memchr is a false positive