@jakobnissen 2025-03-06
@oscardssmith 

I've just taken a look at this. There are a few minor issues in your code:
* Your `findnext` call starts at a pointer, which is not meaningful
* It may search from an invalid index, which will throw an error

I tested the version of your code below, and found it was faster than the current `in(::Char, ::AbstractString)` for ASCII, but slower for non-ASCII.
However, doing `in(c::Char, s::Union{String, SubString{String}}) = @inline(findfirst(==(c), s)) !== nothing` is nearly as fast.

```julia
function foo(c::Char, s::String)
    byte = (reinterpret(UInt32, c) >> 24) % UInt8
    GC.@preserve s begin
        ptr = pointer(s)
        found_ptr = ccall(:memchr, Ptr{UInt8}, (Ptr{UInt8}, Int32, Csize_t), ptr, byte, sizeof(s))
        found_ptr == C_NULL && return false
        byte < 0x80 && return true
        return findnext(isequal(c), s, thisind(s, (found_ptr - ptr) % Int + 1)) !== nothing
    end
end
```