@StefanKarpinski 2025-03-25
Another option: check if the `Char` is valid and if so use `memmem`; if it's invalid, do the slower iteration thing. This works because any valid needle is self-contained; some invalid needles can occur as substrings of other strings.